api_vector_database/requirements.txt
sentence_transformers
PyMuPDF  
fastapi  
pydantic  
pandas  
chromadb  
google-generativeai  
uvicorn  
  
